Rename TestEmailSkill to EmailSkillFake and move to Fakes namespace
      Summary: This commit renames the TestEmailSkill class to EmailSkillFake, to better reflect its purpose as a fake implementation of the email skill for testing. It also moves the class to the Fakes namespace, to avoid confusion with the actual email skill. Additionally, it adds a using directive for the Fakes namespace in PlannerSkillTests.cs and PlanTests.cs, where the EmailSkillFake is used.
